pyproject.toml
[tool.poetry]
name = "agentic ai"
version = "0.1.0"
description = "Agentic AI project managed with Poetry"
authors = ["Abhishek Sharma <2001abhishek17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.61.0"
langchain-openai = "*"
python-dotenv = "^1.0.1"
langchain-community = "*"
browser-use = "*"
langchain = "*"
playwright = "^1.50.0"
gradio = "^5.15.0"
langchain-ollama = "^0.2.3"
langchain-mistralai = "^0.2.6"
json-repair = "^0.35.0"
pyperclip = "^1.9.0"
openai-agents = "^0.0.3"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
